pyproject.toml
[build-system]
build-backend = 'setuptools.build_meta'
requires = ['setuptools>=77.0']

[project]
name = 'pyvista-tui'
version = '0.1.0'
description = 'Interactive 3D mesh visualization in the terminal using PyVista.'
readme = 'README.md'
license = 'MIT'
requires-python = '>=3.10'
authors = [{ name = 'The PyVista Developers' }]
classifiers = [
  'Development Status :: 3 - Alpha',
  'Intended Audience :: Developers',
  'Intended Audience :: Science/Research',
  'Operating System :: MacOS',
  'Operating System :: POSIX :: Linux',
  'Programming Language :: Python :: 3',
  'Programming Language :: Python :: 3.10',
  'Programming Language :: Python :: 3.11',
  'Programming Language :: Python :: 3.12',
  'Programming Language :: Python :: 3.13',
  'Topic :: Scientific/Engineering :: Visualization',
]
keywords = ['3d', 'pyvista', 'terminal', 'tui', 'visualization']
dependencies = [
  'cyclopts>=3.0',
  'Pillow>=10.0',
  'pyvista>=0.43',
  'rich>=13.0',
  'textual>=1.0',
  'textual-image>=0.5',
]

[project.optional-dependencies]
dev = [
  'mypy>=1.10',
  'pytest>=8.0',
  'pytest-asyncio>=1.0',
  'pytest-cov>=5.0',
  'ruff>=0.15',
  'ipython',
]

[project.scripts]
pyvista-tui = 'pyvista_tui.cli:app'

[project.urls]
Homepage = 'https://github.com/pyvista/pyvista-tui'
Issues = 'https://github.com/pyvista/pyvista-tui/issues'
Repository = 'https://github.com/pyvista/pyvista-tui'

[tool.coverage.run]
branch = true
source = ['pyvista_tui']

[tool.coverage.report]
fail_under = 70
exclude_lines = [
  'pragma: no cover',
  'if TYPE_CHECKING:',
  'if __name__',
  '\.\.\.',
]

[tool.pytest.ini_options]
addopts = '-v'
filterwarnings = [
  'error',
  'ignore:pyvista test.*does not yet exist:UserWarning',
  'ignore:Image.Image.getdata is deprecated:DeprecationWarning',
]
testpaths = ['tests']
xfail_strict = true

[tool.ruff]
line-length = 99
src = ['.']

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 75
quote-style = 'single'

[tool.ruff.lint]
ignore = [
  'D107', # Missing docstring in `__init__`
]
extend-select = [
  'A',
  'B',
  'C4',
  'D',
  'E',
  'EM',
  'F',
  'FA',
  'FLY',
  'I',
  'ICN',
  'N',
  'PERF',
  'PGH',
  'PIE',
  'PLC0415',
  'PT',
  'RET',
  'RSE',
  'RUF',
  'SIM',
  'T10',
  'T20',
  'TCH',
  'TID',
  'UP',
  'W',
  'YTT',
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
required-imports = ['from __future__ import annotations']

[tool.ruff.lint.per-file-ignores]
'__init__.py' = ['E402', 'PLC0414']
'scripts/**' = ['D', 'E501', 'FA', 'I002', 'PLC0415', 'T20']
'tests/**' = ['D', 'S101', 'T20', 'TC003']

[tool.ruff.lint.pydocstyle]
convention = 'numpy'

[tool.numpydoc_validation]
checks = [
  'all',
  'ES01', # Not all docstrings need an extended summary
  'EX01', # Examples not required everywhere
  'GL01', # Contradicts numpydoc examples
  'GL02', # Permit blank line after docstring
  'GL03', # Double blank line handling
  'GL08', # Missing docstring
  'PR01', # Undocumented parameters
  'PR02', # Unknown parameters
  'PR03', # Wrong parameters order
  'PR04', # Parameter has no type
  'PR05', # Parameter type ends with "."
  'PR06', # Wrong parameter type
  'PR07', # Parameter has no description
  'PR08', # Parameter desc should start with capital
  'PR09', # Parameter desc should end with "."
  'PR10', # Parameter missing space before colon
  'RT02', # Returns first line type-only
  'SA01', # See Also not required
  'SA04', # See Also needs no descriptions
  'YD01', # Yields not required
]
exclude = [
  '\._.*$', # Ignore private members
]

[tool.mypy]
python_version = '3.10'
ignore_missing_imports = true

[tool.setuptools.package-data]
pyvista_tui = ['py.typed']

[tool.setuptools.packages.find]
where = ['.']
